// src/shared/api/generated/models/speechSettingsResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { SpeechIntegrationResponseDto } from "./speechIntegrationResponseDto";
import type { ProjectSpeechSettingsResponseDto } from "./projectSpeechSettingsResponseDto";

export interface SpeechSettingsResponseDto {
  integration: SpeechIntegrationResponseDto;
  settings: ProjectSpeechSettingsResponseDto;
}